+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+---
+## [v1.2.2] - 2026-01-22 - Fix Mobile Styling
+### Fixed
+- **Mobile Display**
+  - Content is not horizontally scrolling (again yes)
+  - Incorrect padding and styling
+
+### Nakaw Background
+- **Square Background**
+  - From here: https://www.reactbits.dev/backgrounds/squares 
 
 --- 
 ## [v1.2.1] - 2026-01-22 - New Background